+++ lib/bureaucrat/fields/float_field.rb
@@ -8,7 +8,7 @@ module Bureaucrat
       end
 
       def to_object(value)
-        if Validators.empty_value?(value)
+        if value.blank?
           return nil
         end
