--- lib/bureaucrat/widgets.rb
@@ -1,5 +1,0 @@
-# TODO: DateTimeInput
-# TODO: TimeInput
-# TODO: MultiWidget < Widget
-# TODO: SplitDateTimeWidget < MultiWidget
-# TODO: SplitHiddenDateTimeWidget < SplitDateTimeWidget
