+++ spec/quickfields_spec.rb
@@ -1,0 +1,1 @@
+require 'bureaucrat/quickfields'
